+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved `$this->filesystem` from `Bayfront\Bones\Controller` to `Bayfront\Bones\Model` and made it optional.
 - Moved handling CLI and cron jobs to after the app bootstrap file is included and `app.bootstrap` event is executed.
 - Renamed `BONES_ROOT_PATH` and `APP_ROOT_PATH` to `BONES_BASE_PATH` and `APP_BASE_PATH` respectively, as these are more semantically correct. 
+- Updated how the exception handler responds as JSON.
 
 ### Removed